Refactoring to isolate some settings code in Hibernate ORM extension
(cherry picked from commit d61bb8b0889934ed300bf7e5566d831dbae7992b)
