+++ extensions/hibernate-orm/deployment/src/main/java/io/quarkus/hibernate/orm/deployment/HibernateOrmConfig.java
@@ -3,7 +3,6 @@ package io.quarkus.hibernate.orm.deployment;
 import java.util.Map;
 import java.util.Optional;
 
-import io.quarkus.hibernate.orm.runtime.PersistenceUnitUtil;
 import io.quarkus.runtime.annotations.ConfigDocMapKey;
 import io.quarkus.runtime.annotations.ConfigDocSection;
 import io.quarkus.runtime.annotations.ConfigGroup;
@@ -62,13 +61,6 @@ public class HibernateOrmConfig {
                 log.isAnyPropertySet() ||
                 statistics.isPresent() ||
                 metricsEnabled;
-    }
-
-    public static String puPropertyKey(String puName, String radical) {
-        String prefix = PersistenceUnitUtil.isDefaultPersistenceUnit(puName)
-                ? "quarkus.hibernate-orm."
-                : "quarkus.hibernate-orm.\"" + puName + "\".";
-        return prefix + radical;
     }
 
     @ConfigGroup
